umanoar/requirements.txt
Django==2.2.5
djangorestframework==3.10.3
markdown==3.1.1
django-filter==2.2.0
Pillow==6.2.0